requirements.txt
google-api-python-client==2.160.0
python-dotenv==1.0.1
firebase-admin==6.6.0
python-dateutil==2.8.2